+++ src/polyfills/reportValidity.js
@@ -14,7 +14,10 @@ import { get_wrapper } from '../components/wrapper';
 export default function reportValidity(element) {
   /* if this is a <form>, report validity of all child inputs */
   if (element instanceof window.HTMLFormElement) {
-    return get_validated_elements(element).map(reportValidity).every(b=>b);
+    element.__hf_form_validation = true;
+    const form_valid = get_validated_elements(element).map(reportValidity).every(b=>b);
+    delete(element.__hf_form_validation);
+    return form_valid;
   }
 
   /* we copy checkValidity() here, b/c we have to check if the "invalid"