+++ src/validators/required.js
@@ -1,38 +1,34 @@
 'use strict';
 
 
+import { get_radiogroup } from '../tools/get_radiogroup';
+
+
 /**
  * test the required attribute
  */
 export default function(element) {
+  if (element.type === 'radio') {
+    /* the happy (and quick) path for radios: */
+    if (element.hasAttribute('required') && element.checked) {
+      return true;
+    }
+
+    const radiogroup = get_radiogroup(element);
+
+    /* if any radio in the group is required, we need any (not necessarily the
+     * same) radio to be checked */
+    if (radiogroup.some(radio => radio.hasAttribute('required'))) {
+      return radiogroup.some(radio => radio.checked);
+    }
+    /* not required, validation passes */
+    return true;
+  }
+
   if (! element.hasAttribute('required')) {
     /* nothing to do */
     return true;
   }
 
-  /* we don't need get_type() for element.type, because "checkbox" and "radio"
-   * are well supported. */
-  switch (element.type) {
-    case 'checkbox':
-      return element.checked;
-      //break;
-    case 'radio':
-      /* radio inputs have "required" fulfilled, if _any_ other radio
-       * with the same name in this form is checked. */
-      return !! (
-        element.checked ||
-        (
-          element.form &&
-          Array.prototype.filter.call(
-            document.getElementsByName(element.name),
-            radio => radio.name === element.name &&
-                     radio.form === element.form &&
-                     radio.checked
-          ).length > 0
-        )
-      );
-      //break;
-    default:
-      return !! element.value;
-  }
+  return (element.type === 'checkbox')? element.checked : (!! element.value);
 }