+++ test/functional/tests.js
@@ -1,3 +1,4 @@
 window.hyperform = require('../../dist/hyperform.cjs');
 
 require('./test.regressions');
+require('./test.functional');