+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * add `CHANGELOG.md`
 * prevent `is_validation_candidate()` from running twice for each validation
 * re-allow validation of elements without `name` attribute, if asked directly
+* fix validation for partly required radio groups
 
 ## v0.9.14